vpu5: Add software_readable_output mode
When software_readable_output mode is set, the IPMMU and MERAM and bypassed
so that the output data buffers can be used by a software renderer.
This is of course much slower, but it allows for the creation of thumbnails
and other direct access to the frame buffer.

The mode is enabled/disabled by using an OMX extension call to setParameter
with the paramter value being determined by a call to getExtensionIndex with
an index string value "OMX.RE.SoftwareRender".
The setParamter call uses only a OMX_BOOL as data, so the call looks like
OMX_SetParameter(handle, index, (OMX BOOL *) data, sizeof (OMX_BOOL));
where index and data are the value returned from getExtensionIndex and
OMX_TRUE (enable) or OMX_FALSE (disable) respectively.

Additional features/changes may be added to this mode in future