vpu5ha: Change BMI queue to a hash table
Frames may be dropped in the VPU, leaving orphaned BMI data.  Also, frames
can be reordered during decoding, which makes it necessary to have random
access to the BMI data.  The queue structure does not fit well with this
access methodology, so a hash is used instead.  Hash size should be about
equal to the maximum number of stream data that can be buffered.