vpu5: Temporary workaround: Cannot decode after encode
With the current middleware verion (alpha SPF1: 11/11/18), after
encoding one video stream (AVC), the VPU cannot be used to decode
any streams (other encodes are ok).  It looks like the IRQ registers
need to be reset before decoding stars.

This setting is safe to set between decodes as well.

This should really be handled by the middleware and so this patch should
be reverted when a new middleware version is available.