Android.mk: Simplify Android.mk by moving some defines to product definition
Specify the version of the VPU explicitly with it own variable rather than
getting it implicitly from the TARGET_PRODUCT.  This is more flexible and
removes the burden of tracking all of the board configurations from the
Android.mk.  Also, allow the encoder/decoder availablity to be defined
explicitly.  The middleware path is updated to be common across all versions,
but with the ability to be overridden.  The following variables should be
set (if applicable) from the BoardConfig.mk file in the product definition:

PRODUCT_VPU_VERSION
VPU_DECODER_COMPONENT
VPU_ENCODER_COMPONENT
VPU_MIDDLEWARE_PATH

Board specific configurations for mackerel, kota2, and ape5r are left in
for now, but will be removed in future.