vpu5: adjust input buffer handling to return buffers to client sooner
Previously input buffers were only returned to the client after a full
picture's worth of data was recieved and processed.  This poses problems when
a picture is made up of a number of small buffers, instead of one big one.

Adjust the processins so that once all of the NAL units in a buffer have
been processed, the contents are immediately copied out and the buffer
is returned to the client.